service working for resto generation

+++ restaurant/src/main/java/fr/pick_eat/restaurant/utils/CoordCalcul.java
@@ -3,14 +3,14 @@ package fr.pick_eat.restaurant.utils;
 import java.util.List;
 
 public class CoordCalcul {
-    public static List<Double> getAreaFromRadius(double lat, double lon, double radius) {
+    public static List<Float> getAreaFromRadius(float lat, float lon, float radius) {
 //            Get the area from a radius
 //            :param radius: meters
-        double radiusInDegrees = radius / 111000f;
-        double minLat = lat - radiusInDegrees;
-        double maxLat = lat + radiusInDegrees;
-        double minLon = lon - radiusInDegrees;
-        double maxLon = lon + radiusInDegrees;
+        float radiusInDegrees = radius / 111000f;
+        float minLat = lat - radiusInDegrees;
+        float maxLat = lat + radiusInDegrees;
+        float minLon = lon - radiusInDegrees;
+        float maxLon = lon + radiusInDegrees;
         return List.of(minLat, maxLat, minLon, maxLon);
     }
 }
